refactor: excluindo arquivos individuais das telas modAluno para unificar em um único arquivo.
