--- src/modAluno.h
@@ -1,6 +1,0 @@
-#ifndef MOD_ALUNO_H
-#define MOD_ALUNO_H
-
-char moduloAluno(void);
-
-#endif
