--- src/ui/aluno/atualizarAluno.c
@@ -1,21 +1,0 @@
-#include <stdio.h>
-
-char telaAtualizarAluno(void) {
-
-    char op;
-
-    printf("\n");
-    printf("====================\n");
-    printf(" ATUALIZAR ALUNO\n");
-    printf("====================\n");
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-    scanf("%c", &op);
-    getchar();
-
-    
-
-    return op;
-
-    }
