--- src/ui/aluno/atualizarAluno.h
@@ -1,6 +1,0 @@
-#ifndef ATUALIZAR_ALUNO_H
-#define ATUALIZAR_ALUNO_H
-
-char telaAtualizarAluno(void);
-
-#endif
