--- src/ui/aluno/cadastrarAluno.h
@@ -1,6 +1,0 @@
-#ifndef CADASTRAR_ALUNO_H
-#define CADASTRAR_ALUNO_H
-
-char telaCadastrarAluno(void);
-
-#endif
