--- src/ui/aluno/excluirAluno.h
@@ -1,6 +1,0 @@
-#ifndef EXCLUIR_ALUNO_H
-#define EXCLUIR_ALUNO_H
-
-char telaExcluirAluno(void);
-
-#endif
