--- src/ui/aluno/telaAluno.h
@@ -1,6 +1,0 @@
-#ifndef TELA_ALUNO_H
-#define TELA_ALUNO_H
-
-char telaAluno(void);
-
-#endif
