--- src/ui/aluno/visualizarAluno.h
@@ -1,6 +1,0 @@
-#ifndef VISUALIZAR_ALUNO_H
-#define VISUALIZAR_ALUNO_H
-
-char telaVisualizarAluno(void);
-
-#endif
