+++ src/ui/telasAluno.h
@@ -1,0 +1,11 @@
+#ifndef TELASALUNO_H
+#define TELASALUNO_H
+
+char moduloAluno(void);
+char telaAluno(void);
+char telaCadastrarAluno(void);
+char telaVisualizarAluno(void);
+char telaAtualizarAluno(void);
+char telaExcluirAluno(void);
+
+#endif
